+++ src/minoversikt/minoversikt-side.tsx
@@ -113,7 +113,7 @@ export default function MinoversiktSide() {
     return (
         <DocumentTitle title="Min oversikt">
             <div className="side-storrelse" id={`side-storrelse_${id}`} data-testid={`side-storrelse_${id}`}>
-                <ToppMeny erPaloggetVeileder={!visesAnnenVeiledersPortefolje} />
+                <ToppMeny erPaloggetVeileder={!visesAnnenVeiledersPortefolje} oversiktType={oversiktType} />
                 <AlertstripeTekniskeProblemer />
                 <Innholdslaster avhengigheter={[statustall]}>
                     <MinOversiktWrapper
