+++ src/minoversikt/moteplan/moteplan.less
@@ -1,5 +1,5 @@
 .veilarbportefoljeflatefs {
-    .motekalender_knapp {
+    .moteplan_knapp {
         height: fit-content;
         margin-top: 0.3125rem !important;
     }
@@ -10,13 +10,13 @@
 
     .moteplan_tittel {
         margin-bottom: 0;
-        margin-left: 0.75rem;
+        margin-left: 8px;
     }
 
     .moteplan_tabell_tittelrad th {
         margin-bottom: 0;
         padding-bottom: 0 !important;
-        font-size: 0.875rem;
+        font-size: 1rem;
         font-weight: normal;
         font-variant: all-small-caps;
     }
