+++ src/veiledere/veiledere-side.tsx
@@ -59,7 +59,7 @@ function VeiledereSide() {
                 id={`side-storrelse_${id}`}
                 data-testid={`side-storrelse_${id}`}
             >
-                <ToppMeny />
+                <ToppMeny oversiktType={oversiktType} />
                 <AlertstripeTekniskeProblemer />
                 <Innholdslaster avhengigheter={[statustall]}>
                     <div
